Update user group schema tests
