+++ tests/schemas/test_user_group_invalid_schema_cases.py
@@ -1,0 +1,6 @@
+from typing import Literal
+
+
+class CaseInvalidAttr:
+    def case_attr(self) -> tuple[Literal["name"], None]:
+        return "name", None
